A few fixes.

+++ src/mcmc_chains.jl
@@ -37,7 +37,7 @@ function MCMCChains.ess(
     chn::MCMCChains.AbstractChains;
     showall::Bool=false,
     sections::Union{Symbol, Vector{Symbol}}=Symbol[:parameters],
-    maxlag = 250
+    maxlag::Int = 250
 )
     param = showall ? names(chn) : names(chn, sections)
     n_chain_orig = size(chn, 3)
